code/requirements.txt
pynmcli
python-nmap
requests